858: Refactor to split meals system calendar specs
