+++ spec/support/contexts/calendars/meals_system_calendars_context.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+shared_context "meals system calendars" do
+  let(:community) { Defaults.community }
+  let(:communityB) { create(:community) }
+  let(:user) { create(:user, community: community) }
+  let(:full_range) { (Time.current - 2.days)..(Time.current + 4.days) }
+  let!(:meal1) do
+    create(:meal, head_cook: user, served_at: Time.current + 1.day)
+  end
+  let!(:meal2) do
+    create(:meal, :with_menu, title: "Meal2", served_at: Time.current + 2.days)
+  end
+  let!(:meal3) do
+    create(:meal, :with_menu, title: "Other Cmty Meal", community: communityB,
+                              served_at: Time.current + 3.days,
+                              communities: [meal1.community, communityB])
+  end
+  let!(:meal4) do
+    create(:meal, :with_menu, title: "Other Cmty Meal 2", community: communityB,
+                              served_at: Time.current + 4.days,
+                              communities: [meal1.community, communityB])
+  end
+  let!(:cancelled_meal) do
+    create(:meal, :cancelled, served_at: meal1.served_at)
+  end
+  let!(:signup1) do
+    create(:meal_signup, meal: meal1, household: user.household, comments: "Foo\nBar", diner_counts: [2])
+  end
+  let!(:signup2) do
+    create(:meal_signup, meal: cancelled_meal, household: user.household, diner_counts: [2])
+  end
+  let!(:signup3) do
+    create(:meal_signup, meal: meal3, household: user.household, comments: "Foo\nBar", diner_counts: [2])
+  end
+end
